requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
GitPython>=3.1.40
